Updates. Still WIP. Also: Updated copyright line.

+++ app/src/main/java/nl/joozd/logbookapp/data/repository/helpers/flightExtensions.kt
@@ -1,3 +1,22 @@
+/*
+ *  JoozdLog Pilot's Logbook
+ *  Copyright (c) 2020 Joost Welle
+ *
+ *      This program is free software: you can redistribute it and/or modify
+ *      it under the terms of the GNU Affero General Public License as
+ *      published by the Free Software Foundation, either version 3 of the
+ *      License, or (at your option) any later version.
+ *
+ *      This program is distributed in the hope that it will be useful,
+ *      but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *      MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *      GNU Affero General Public License for more details.
+ *
+ *      You should have received a copy of the GNU Affero General Public License
+ *      along with this program.  If not, see https://www.gnu.org/licenses
+ *
+ */
+
 package nl.joozd.logbookapp.data.repository.helpers
 
 import nl.joozd.logbookapp.extensions.toInt
@@ -10,5 +29,5 @@ const val PLANNING_MARGIN = 300 // seconds = 5 minutes. Flights saved with timeI
 
 fun Flight.prepareForSave(): Flight{
     val now = Instant.now().epochSecond
-    return this.copy(isPlanned = (timeIn > now + PLANNING_MARGIN).toInt(), timeStamp = TimestampMaker.nowForSycPurposes)
+    return this.copy(isPlanned = (timeIn > now + PLANNING_MARGIN), timeStamp = TimestampMaker.nowForSycPurposes)
 }
